refactor/split functions to a different files
