+++ src/classes/UserCart.ts
@@ -11,7 +11,9 @@ import { EconItem } from 'steam-tradeoffer-manager';
 import { CurrencyObject, CurrencyObjectWithWeapons, Currency } from '../types/TeamFortress2';
 import { UnknownDictionary } from '../types/common';
 
-import { craftAll, uncraftAll, noiseMakerSKU, noiseMakerNames, strangeParts, spMore1Keys } from '../lib/data';
+import { craftAll, uncraftAll, noiseMakerSKU } from '../lib/data';
+import { pure } from '../lib/tools/pure';
+import { checkUses, checkHighValue } from '../lib/tools/check';
 // import { parseEconItem } from 'tf2-item-format';
 
 import log from '../lib/logger';
@@ -467,199 +469,33 @@ class UserCart extends Cart {
                 process.env.DISABLE_CHECK_USES_DUELING_MINI_GAME === 'false' ||
                 process.env.DISABLE_CHECK_USES_NOISE_MAKER === 'false'
             ) {
-                let hasNot5Uses = false;
-                let hasNot25Uses = false;
-
                 if (sku === '241;6' || noiseMakerSKU.includes(sku)) {
-                    fetched.forEach(item => {
-                        const isDuelingMiniGame = item.market_hash_name === 'Dueling Mini-Game';
-                        const isNoiseMaker = noiseMakerNames.some(name => {
-                            return item.market_hash_name.includes(name);
-                        });
+                    const yes: {
+                        isNot5Uses: boolean;
+                        isNot25Uses: boolean;
+                    } = checkUses(offer, fetched, this.bot);
 
-                        if (isDuelingMiniGame && process.env.DISABLE_CHECK_USES_DUELING_MINI_GAME === 'false') {
-                            for (let i = 0; i < item.descriptions.length; i++) {
-                                const descriptionValue = item.descriptions[i].value;
-                                const descriptionColor = item.descriptions[i].color;
+                    if (yes.isNot5Uses) {
+                        return Promise.reject(
+                            'One of your Dueling Mini-Game is not 5 Uses. Please make sure you only have 5 Uses in your inventory or send me an offer with the one that has 5 Uses instead'
+                        );
+                    }
 
-                                if (
-                                    !descriptionValue.includes('This is a limited use item. Uses: 5') &&
-                                    descriptionColor === '00a000'
-                                ) {
-                                    hasNot5Uses = true;
-                                    offer.log('info', 'contains Dueling Mini-Game that is not 5 uses, declining...');
-                                    break;
-                                }
-                            }
-                        } else if (isNoiseMaker && process.env.DISABLE_CHECK_USES_NOISE_MAKER === 'false') {
-                            for (let i = 0; i < item.descriptions.length; i++) {
-                                const descriptionValue = item.descriptions[i].value;
-                                const descriptionColor = item.descriptions[i].color;
-
-                                if (
-                                    !descriptionValue.includes('This is a limited use item. Uses: 25') &&
-                                    descriptionColor === '00a000'
-                                ) {
-                                    hasNot25Uses = true;
-                                    offer.log('info', `${item.market_hash_name} (${item.assetid}) is not 25 uses.`);
-                                    break;
-                                }
-                            }
-                        }
-                    });
-                }
-
-                if (hasNot5Uses) {
-                    return Promise.reject(
-                        'One of your Dueling Mini-Game is not 5 Uses. Please make sure you only have 5 Uses in your inventory or send me an offer with the one that has 5 Uses instead'
-                    );
-                }
-
-                if (hasNot25Uses) {
-                    return Promise.reject(
-                        'One of your Noise Maker in your inventory is not 25 Uses. Please make sure you only have 25 Uses in your inventory or send me an offer with the one that has 25 Uses instead'
-                    );
+                    if (yes.isNot25Uses) {
+                        return Promise.reject(
+                            'One of your Noise Maker in your inventory is not 25 Uses. Please make sure you only have 25 Uses in your inventory or send me an offer with the one that has 25 Uses instead'
+                        );
+                    }
                 }
             }
 
             const toMention = (this.bot.handler as MyHandler).getToMention();
             const highValuedTheir: {
+                has: boolean;
                 skus: string[];
                 names: string[];
                 isMention: boolean;
-            } = {
-                skus: [],
-                names: [],
-                isMention: false
-            };
-
-            const isEnabledDiscordWebhook =
-                process.env.DISABLE_DISCORD_WEBHOOK_TRADE_SUMMARY === 'false' &&
-                process.env.DISCORD_WEBHOOK_TRADE_SUMMARY_URL;
-
-            fetched.forEach(item => {
-                // const parsed = parseEconItem(
-                //     {
-                //         ...item,
-                //         tradable: item.tradable ? 1 : 0,
-                //         commodity: item.commodity ? 1 : 0,
-                //         marketable: item.marketable ? 1 : 0,
-                //         amount: item.amount + ''
-                //     },
-                //     true,
-                //     true
-                // );
-
-                const itemSKU = item.getSKU(this.bot.schema);
-
-                if (sku === itemSKU) {
-                    // let hasSpelled = false;
-                    // if (parsed.spells.length > 0) {
-                    //     hasSpelled = true;
-                    // }
-
-                    // let hasStrangeParts = false;
-                    // if (parsed.parts.length > 0) {
-                    //     hasStrangeParts = true;
-                    // }
-
-                    let hasSpells = false;
-                    let hasStrangeParts = false;
-                    let hasKillstreaker = false;
-                    let hasSheen = false;
-
-                    const spellNames: string[] = [];
-                    const partsNames: string[] = [];
-                    const killstreakerName: string[] = [];
-                    const sheenName: string[] = [];
-
-                    for (let i = 0; i < item.descriptions.length; i++) {
-                        const desc = item.descriptions[i].value;
-                        const parts = item.descriptions[i].value
-                            .replace('(', '')
-                            .replace(/: \d+\)/g, '')
-                            .trim();
-
-                        const color = item.descriptions[i].color;
-                        const strangePartNames = Object.keys(strangeParts);
-
-                        if (
-                            desc.startsWith('Halloween:') &&
-                            desc.endsWith('(spell only active during event)') &&
-                            color === '7ea9d1'
-                        ) {
-                            hasSpells = true;
-                            highValuedTheir.isMention = true;
-                            const spellName = desc.substring(10, desc.length - 32).trim();
-                            spellNames.push(spellName);
-                        } else if (
-                            (parts === 'Kills' || parts === 'Assists'
-                                ? item.type.includes('Strange') && item.type.includes('Points Scored')
-                                : strangePartNames.includes(parts)) &&
-                            color === '756b5e'
-                        ) {
-                            hasStrangeParts = true;
-                            if (Object.keys(spMore1Keys).includes(parts)) {
-                                highValuedTheir.isMention = true;
-                                partsNames.push(parts + ' (>🔑)');
-                            } else {
-                                partsNames.push(parts);
-                            }
-                        } else if (desc.startsWith('Killstreaker: ') && color === '7ea9d1') {
-                            const extractedName = desc.replace('Killstreaker: ', '').trim();
-                            hasKillstreaker = true;
-                            if (toMention.killstreakers.includes(extractedName.toLowerCase())) {
-                                highValuedTheir.isMention = true;
-                                killstreakerName.push(extractedName + ' (🌟)');
-                            } else {
-                                killstreakerName.push(extractedName);
-                            }
-                        } else if (desc.startsWith('Sheen: ') && color === '7ea9d1') {
-                            const extractedName = desc.replace('Sheen: ', '').trim();
-                            hasSheen = true;
-                            if (toMention.sheens.includes(extractedName.toLowerCase())) {
-                                highValuedTheir.isMention = true;
-                                sheenName.push(extractedName + ' (🌟)');
-                            } else {
-                                sheenName.push(extractedName);
-                            }
-                        }
-                    }
-
-                    if (hasSpells || hasStrangeParts || hasKillstreaker || hasSheen) {
-                        highValuedTheir.skus.push(itemSKU);
-                        const itemObj = SKU.fromString(itemSKU);
-                        const itemName =
-                            itemObj.quality === 5 ? this.bot.schema.getName(itemObj, false) : item.market_hash_name;
-
-                        let itemDescriptions = '';
-
-                        if (hasSpells) {
-                            itemDescriptions += '\n🎃 Spells: ' + spellNames.join(' + ');
-                        }
-
-                        if (hasStrangeParts) {
-                            itemDescriptions += '\n🎰 Parts: ' + partsNames.join(' + ');
-                        }
-
-                        if (hasKillstreaker) {
-                            itemDescriptions += '\n🔥 Killstreker: ' + killstreakerName.join(' + ');
-                        }
-
-                        if (hasSheen) {
-                            itemDescriptions += '\n✨ Sheen: ' + sheenName.join(' + ');
-                        }
-
-                        log.debug('info', `${itemName} (${item.assetid})${itemDescriptions}`);
-
-                        if (isEnabledDiscordWebhook) {
-                            highValuedTheir.names.push(`_${itemName}_${itemDescriptions}`);
-                        } else {
-                            highValuedTheir.names.push(`${itemName}${itemDescriptions}`);
-                        }
-                    }
-                }
-            });
+            } = checkHighValue(fetched, toMention.sheens, toMention.killstreakers, this.bot);
 
             offer.data('highValue', highValuedTheir);
 
@@ -731,7 +567,7 @@ class UserCart extends Cart {
 
         // We now know who the buyer is, now get their inventory
         const buyerInventory = isBuyer ? this.bot.inventoryManager.getInventory() : theirInventory;
-        const pureStock = (this.bot.handler as MyHandler).pureStock();
+        const pureStock = pure(this.bot);
 
         if (this.bot.inventoryManager.amountCanAfford(this.canUseKeys(), currencies, buyerInventory) < 1) {
             // Buyer can't afford the items
@@ -2013,199 +1849,33 @@ class UserCart extends Cart {
                 process.env.DISABLE_CHECK_USES_DUELING_MINI_GAME === 'false' ||
                 process.env.DISABLE_CHECK_USES_NOISE_MAKER === 'false'
             ) {
-                let hasNot5Uses = false;
-                let hasNot25Uses = false;
-
                 if (sku === '241;6' || noiseMakerSKU.includes(sku)) {
-                    fetched.forEach(item => {
-                        const isDuelingMiniGame = item.market_hash_name === 'Dueling Mini-Game';
-                        const isNoiseMaker = noiseMakerNames.some(name => {
-                            return item.market_hash_name.includes(name);
-                        });
+                    const yes: {
+                        isNot5Uses: boolean;
+                        isNot25Uses: boolean;
+                    } = checkUses(offer, fetched, this.bot);
 
-                        if (isDuelingMiniGame && process.env.DISABLE_CHECK_USES_DUELING_MINI_GAME === 'false') {
-                            for (let i = 0; i < item.descriptions.length; i++) {
-                                const descriptionValue = item.descriptions[i].value;
-                                const descriptionColor = item.descriptions[i].color;
+                    if (yes.isNot5Uses) {
+                        return Promise.reject(
+                            'One of your Dueling Mini-Game is not 5 Uses. Please make sure you only have 5 Uses in your inventory or send me an offer with the one that has 5 Uses instead'
+                        );
+                    }
 
-                                if (
-                                    !descriptionValue.includes('This is a limited use item. Uses: 5') &&
-                                    descriptionColor === '00a000'
-                                ) {
-                                    hasNot5Uses = true;
-                                    offer.log('info', 'contains Dueling Mini-Game that is not 5 uses, declining...');
-                                    break;
-                                }
-                            }
-                        } else if (isNoiseMaker && process.env.DISABLE_CHECK_USES_NOISE_MAKER === 'false') {
-                            for (let i = 0; i < item.descriptions.length; i++) {
-                                const descriptionValue = item.descriptions[i].value;
-                                const descriptionColor = item.descriptions[i].color;
-
-                                if (
-                                    !descriptionValue.includes('This is a limited use item. Uses: 25') &&
-                                    descriptionColor === '00a000'
-                                ) {
-                                    hasNot25Uses = true;
-                                    offer.log('info', `${item.market_hash_name} (${item.assetid}) is not 25 uses.`);
-                                    break;
-                                }
-                            }
-                        }
-                    });
-                }
-
-                if (hasNot5Uses) {
-                    return Promise.reject(
-                        'One of your Dueling Mini-Game is not 5 Uses. Please make sure you only have 5 Uses in your inventory or send me an offer with the one that has 5 Uses instead'
-                    );
-                }
-
-                if (hasNot25Uses) {
-                    return Promise.reject(
-                        'One of your Noise Maker in your inventory is not 25 Uses. Please make sure you only have 25 Uses in your inventory or send me an offer with the one that has 25 Uses instead'
-                    );
+                    if (yes.isNot25Uses) {
+                        return Promise.reject(
+                            'One of your Noise Maker in your inventory is not 25 Uses. Please make sure you only have 25 Uses in your inventory or send me an offer with the one that has 25 Uses instead'
+                        );
+                    }
                 }
             }
 
             const toMention = (this.bot.handler as MyHandler).getToMention();
             const highValuedTheir: {
+                has: boolean;
                 skus: string[];
                 names: string[];
                 isMention: boolean;
-            } = {
-                skus: [],
-                names: [],
-                isMention: false
-            };
-
-            const isEnabledDiscordWebhook =
-                process.env.DISABLE_DISCORD_WEBHOOK_TRADE_SUMMARY === 'false' &&
-                process.env.DISCORD_WEBHOOK_TRADE_SUMMARY_URL;
-
-            fetched.forEach(item => {
-                // const parsed = parseEconItem(
-                //     {
-                //         ...item,
-                //         tradable: item.tradable ? 1 : 0,
-                //         commodity: item.commodity ? 1 : 0,
-                //         marketable: item.marketable ? 1 : 0,
-                //         amount: item.amount + ''
-                //     },
-                //     true,
-                //     true
-                // );
-
-                const itemSKU = item.getSKU(this.bot.schema);
-
-                if (sku === itemSKU) {
-                    // let hasSpelled = false;
-                    // if (parsed.spells.length > 0) {
-                    //     hasSpelled = true;
-                    // }
-
-                    // let hasStrangeParts = false;
-                    // if (parsed.parts.length > 0) {
-                    //     hasStrangeParts = true;
-                    // }
-
-                    let hasSpells = false;
-                    let hasStrangeParts = false;
-                    let hasKillstreaker = false;
-                    let hasSheen = false;
-
-                    const spellNames: string[] = [];
-                    const partsNames: string[] = [];
-                    const killstreakerName: string[] = [];
-                    const sheenName: string[] = [];
-
-                    for (let i = 0; i < item.descriptions.length; i++) {
-                        const desc = item.descriptions[i].value;
-                        const parts = item.descriptions[i].value
-                            .replace('(', '')
-                            .replace(/: \d+\)/g, '')
-                            .trim();
-
-                        const color = item.descriptions[i].color;
-                        const strangePartNames = Object.keys(strangeParts);
-
-                        if (
-                            desc.startsWith('Halloween:') &&
-                            desc.endsWith('(spell only active during event)') &&
-                            color === '7ea9d1'
-                        ) {
-                            hasSpells = true;
-                            highValuedTheir.isMention = true;
-                            const spellName = desc.substring(10, desc.length - 32).trim();
-                            spellNames.push(spellName);
-                        } else if (
-                            (parts === 'Kills' || parts === 'Assists'
-                                ? item.type.includes('Strange') && item.type.includes('Points Scored')
-                                : strangePartNames.includes(parts)) &&
-                            color === '756b5e'
-                        ) {
-                            hasStrangeParts = true;
-                            if (Object.keys(spMore1Keys).includes(parts)) {
-                                highValuedTheir.isMention = true;
-                                partsNames.push(parts + ' (>🔑)');
-                            } else {
-                                partsNames.push(parts);
-                            }
-                        } else if (desc.startsWith('Killstreaker: ') && color === '7ea9d1') {
-                            const extractedName = desc.replace('Killstreaker: ', '').trim();
-                            hasKillstreaker = true;
-                            if (toMention.killstreakers.includes(extractedName.toLowerCase())) {
-                                highValuedTheir.isMention = true;
-                                killstreakerName.push(extractedName + ' (🌟)');
-                            } else {
-                                killstreakerName.push(extractedName);
-                            }
-                        } else if (desc.startsWith('Sheen: ') && color === '7ea9d1') {
-                            const extractedName = desc.replace('Sheen: ', '').trim();
-                            hasSheen = true;
-                            if (toMention.sheens.includes(extractedName.toLowerCase())) {
-                                highValuedTheir.isMention = true;
-                                sheenName.push(extractedName + ' (🌟)');
-                            } else {
-                                sheenName.push(extractedName);
-                            }
-                        }
-                    }
-
-                    if (hasSpells || hasStrangeParts || hasKillstreaker || hasSheen) {
-                        highValuedTheir.skus.push(itemSKU);
-                        const itemObj = SKU.fromString(itemSKU);
-                        const itemName =
-                            itemObj.quality === 5 ? this.bot.schema.getName(itemObj, false) : item.market_hash_name;
-
-                        let itemDescriptions = '';
-
-                        if (hasSpells) {
-                            itemDescriptions += '\n🎃 Spells: ' + spellNames.join(' + ');
-                        }
-
-                        if (hasStrangeParts) {
-                            itemDescriptions += '\n🎰 Parts: ' + partsNames.join(' + ');
-                        }
-
-                        if (hasKillstreaker) {
-                            itemDescriptions += '\n🔥 Killstreker: ' + killstreakerName.join(' + ');
-                        }
-
-                        if (hasSheen) {
-                            itemDescriptions += '\n✨ Sheen: ' + sheenName.join(' + ');
-                        }
-
-                        log.debug('info', `${itemName} (${item.assetid})${itemDescriptions}`);
-
-                        if (isEnabledDiscordWebhook) {
-                            highValuedTheir.names.push(`_${itemName}_${itemDescriptions}`);
-                        } else {
-                            highValuedTheir.names.push(`${itemName}${itemDescriptions}`);
-                        }
-                    }
-                }
-            });
+            } = checkHighValue(fetched, toMention.sheens, toMention.killstreakers, this.bot);
 
             offer.data('highValue', highValuedTheir);
 
@@ -2277,7 +1947,7 @@ class UserCart extends Cart {
 
         // We now know who the buyer is, now get their inventory
         const buyerInventory = isBuyer ? this.bot.inventoryManager.getInventory() : theirInventory;
-        const pureStock = (this.bot.handler as MyHandler).pureStock();
+        const pureStock = pure(this.bot);
 
         if (this.bot.inventoryManager.amountCanAfford(this.canUseKeysWithWeapons(), currencies, buyerInventory) < 1) {
             // Buyer can't afford the items
