+++ src/lib/data.ts
@@ -738,84 +738,85 @@ export const crates = {
         'Salvaged Mann Co. Supply Crate #30': 30,
         'Salvaged Mann Co. Supply Crate #40': 40,
         'Salvaged Mann Co. Supply Crate #50': 50
-    },
-    isOther: {
-        5048: 6, // Festive Winter Crate #6
-        5066: 22, // Refreshing Summer Cooler #22
-        5070: 35, // Naughty Winter Crate #35
-        5071: 36, // Nice Winter Crate #36
-        5078: 46, // Scorched Crate #46
-        5080: 48, // Fall Crate #48
-        5627: 51, // Eerie Crate #51
-        5629: 52, // Naughty Winter Crate 2012 #52
-        5630: 53, // Nice Winter Crate 2012 #53
-        5635: 58, // Robo Community Crate #58
-        5660: 60, // Select Reserve Mann Co. Supply Crate #60
-        5640: 61, // Summer Appetizer Crate #61
-        5642: 62, // Red Summer 2013 Cooler #62
-        5644: 63, // Orange Summer 2013 Cooler #63
-        5646: 64, // Yellow Summer 2013 Cooler #64
-        5648: 65, // Green Summer 2013 Cooler #65
-        5650: 66, // Aqua Summer 2013 Cooler #66
-        5652: 67, // Blue Summer 2013 Cooler #67
-        5654: 68, // Brown Summer 2013 Cooler #68
-        5656: 69, // Black Summer 2013 Cooler #69 // No #70
-        5708: 72, // Fall 2013 Acorns Crate #72
-        5709: 73, // Fall 2013 Gourd Crate #73
-        5712: 74, // Spooky Crate #74
-        5714: 78, // Naughty Winter Crate 2013 #78
-        5715: 79, // Nice Winter Crate 2013 #79 // No #80
-        5719: 81, // Mann Co. Strongbox #81
-        5734: 82, // Mann Co. Munition Crate #82
-        5735: 83, // Mann Co. Munition Crate #83
-        5742: 84, // Mann Co. Munition Crate #84
-        5752: 85, // Mann Co. Munition Crate #85
-        5761: 86, // Limited Late Summer Crate #86
-        5774: 87, // End of the Line Community Crate #87
-        5789: 88, // Naughty Winter Crate 2014 #88
-        5790: 89, // Nice Winter Crate 2014 #89
-        5781: 90, // Mann Co. Munition Crate #90
-        5802: 91, // Mann Co. Munition Crate #91
-        5803: 92, // Mann Co. Munition Crate #92
-        5806: 93, // The Concealed Killer Weapons Case #93
-        5807: 94, // The Powerhouse Weapons Case #94
-        5817: 95, // Gun Mettle Cosmetic Case #95
-        5822: 96, // Quarantined Collection Case #96
-        5823: 97, // Confidential Collection Case #97
-        5828: 98, // Gargoyle Case #98
-        5831: 99, // Pyroland Weapons Case #99
-        5832: 100, // Warbird Weapons Case #100
-        5842: 101, // Tough Break Cosmetic Case #101
-        5849: 102, // Mayflower Cosmetic Case #102
-        5859: 103, // Mann Co. Munition Crate #103
-        5861: 104, // Creepy Crawly Case #104
-        5865: 105, // Unlocked Winter 2016 Cosmetic Case #105
-        5867: 106, // Rainy Day Cosmetic Case #106
-        5871: 107, // Abominable Cosmetic Case #107
-        5875: 108, // Unleash the Beast Cosmetic Case #108
-        5883: 109, // Jungle Jackpot War Paint Case #109
-        5885: 110, // Infernal Reward War Paint Case #110
-        18000: 111, // 'Decorated War Hero' War Paint\nCivilian Grade Keyless Case
-        18001: 112, // 'Decorated War Hero' War Paint\nFreelance Grade Keyless Case
-        18002: 113, // 'Decorated War Hero' War Paint\nMercenary Grade Keyless Case
-        18003: 114, // 'Contract Campaigner' War Paint\nCivilian Grade Keyless Case
-        18004: 115, // 'Contract Campaigner' War Paint\nFreelance Grade Keyless Case
-        18005: 116, // 'Contract Campaigner' War Paint\nMercenary Grade Keyless Case
-        5888: 117, // Winter 2017 Cosmetic Case #117
-        5890: 118, // Winter 2017 War Paint Case #118
-        5893: 119, // Blue Moon Cosmetic Case #119
-        5894: 120, // Violet Vermin Case #120
-        5897: 121, // Scream Fortress X War Paint Case #121
-        5902: 122, // Winter 2018 Cosmetic Case #122
-        5904: 123, // Summer 2019 Cosmetic Case #123
-        5905: 124, // Spooky Spoils Case #124
-        5909: 125, // Winter 2019 Cosmetic Case #125
-        5912: 126, // Winter 2019 War Paint Case #126
-        5914: 127, // Summer 2020 Cosmetic Case #127
-        5915: 128, // Wicked Windfall Case #128
-        5918: 129 // Scream Fortress XII War Paint Case #129
     }
 };
+
+// isOther: {
+//     5048: 6, // Festive Winter Crate #6
+//     5066: 22, // Refreshing Summer Cooler #22
+//     5070: 35, // Naughty Winter Crate #35
+//     5071: 36, // Nice Winter Crate #36
+//     5078: 46, // Scorched Crate #46
+//     5080: 48, // Fall Crate #48
+//     5627: 51, // Eerie Crate #51
+//     5629: 52, // Naughty Winter Crate 2012 #52
+//     5630: 53, // Nice Winter Crate 2012 #53
+//     5635: 58, // Robo Community Crate #58
+//     5660: 60, // Select Reserve Mann Co. Supply Crate #60
+//     5640: 61, // Summer Appetizer Crate #61
+//     5642: 62, // Red Summer 2013 Cooler #62
+//     5644: 63, // Orange Summer 2013 Cooler #63
+//     5646: 64, // Yellow Summer 2013 Cooler #64
+//     5648: 65, // Green Summer 2013 Cooler #65
+//     5650: 66, // Aqua Summer 2013 Cooler #66
+//     5652: 67, // Blue Summer 2013 Cooler #67
+//     5654: 68, // Brown Summer 2013 Cooler #68
+//     5656: 69, // Black Summer 2013 Cooler #69 // No #70
+//     5708: 72, // Fall 2013 Acorns Crate #72
+//     5709: 73, // Fall 2013 Gourd Crate #73
+//     5712: 74, // Spooky Crate #74
+//     5714: 78, // Naughty Winter Crate 2013 #78
+//     5715: 79, // Nice Winter Crate 2013 #79 // No #80
+//     5719: 81, // Mann Co. Strongbox #81
+//     5734: 82, // Mann Co. Munition Crate #82
+//     5735: 83, // Mann Co. Munition Crate #83
+//     5742: 84, // Mann Co. Munition Crate #84
+//     5752: 85, // Mann Co. Munition Crate #85
+//     5761: 86, // Limited Late Summer Crate #86
+//     5774: 87, // End of the Line Community Crate #87
+//     5789: 88, // Naughty Winter Crate 2014 #88
+//     5790: 89, // Nice Winter Crate 2014 #89
+//     5781: 90, // Mann Co. Munition Crate #90
+//     5802: 91, // Mann Co. Munition Crate #91
+//     5803: 92, // Mann Co. Munition Crate #92
+//     5806: 93, // The Concealed Killer Weapons Case #93
+//     5807: 94, // The Powerhouse Weapons Case #94
+//     5817: 95, // Gun Mettle Cosmetic Case #95
+//     5822: 96, // Quarantined Collection Case #96
+//     5823: 97, // Confidential Collection Case #97
+//     5828: 98, // Gargoyle Case #98
+//     5831: 99, // Pyroland Weapons Case #99
+//     5832: 100, // Warbird Weapons Case #100
+//     5842: 101, // Tough Break Cosmetic Case #101
+//     5849: 102, // Mayflower Cosmetic Case #102
+//     5859: 103, // Mann Co. Munition Crate #103
+//     5861: 104, // Creepy Crawly Case #104
+//     5865: 105, // Unlocked Winter 2016 Cosmetic Case #105
+//     5867: 106, // Rainy Day Cosmetic Case #106
+//     5871: 107, // Abominable Cosmetic Case #107
+//     5875: 108, // Unleash the Beast Cosmetic Case #108
+//     5883: 109, // Jungle Jackpot War Paint Case #109
+//     5885: 110, // Infernal Reward War Paint Case #110
+//     18000: 111, // 'Decorated War Hero' War Paint\nCivilian Grade Keyless Case
+//     18001: 112, // 'Decorated War Hero' War Paint\nFreelance Grade Keyless Case
+//     18002: 113, // 'Decorated War Hero' War Paint\nMercenary Grade Keyless Case
+//     18003: 114, // 'Contract Campaigner' War Paint\nCivilian Grade Keyless Case
+//     18004: 115, // 'Contract Campaigner' War Paint\nFreelance Grade Keyless Case
+//     18005: 116, // 'Contract Campaigner' War Paint\nMercenary Grade Keyless Case
+//     5888: 117, // Winter 2017 Cosmetic Case #117
+//     5890: 118, // Winter 2017 War Paint Case #118
+//     5893: 119, // Blue Moon Cosmetic Case #119
+//     5894: 120, // Violet Vermin Case #120
+//     5897: 121, // Scream Fortress X War Paint Case #121
+//     5902: 122, // Winter 2018 Cosmetic Case #122
+//     5904: 123, // Summer 2019 Cosmetic Case #123
+//     5905: 124, // Spooky Spoils Case #124
+//     5909: 125, // Winter 2019 Cosmetic Case #125
+//     5912: 126, // Winter 2019 War Paint Case #126
+//     5914: 127, // Summer 2020 Cosmetic Case #127
+//     5915: 128, // Wicked Windfall Case #128
+//     5918: 129 // Scream Fortress XII War Paint Case #129
+// }
 
 export const australiumImageURL = {
     // Australium Ambassador
