+++ src/lib/extend/offer/summarizeWithLink.ts
@@ -5,6 +5,7 @@ import SchemaManager from 'tf2-schema-2';
 
 import Currencies from 'tf2-currencies';
 import SKU from 'tf2-sku-2';
+import { replaceItemName } from '../../tools/replace';
 
 export = function(schema: SchemaManager.Schema): string {
     // @ts-ignore
@@ -52,12 +53,7 @@ function summarizeItemsWithLink(dict: UnknownDictionary<number>, schema: SchemaM
         }
 
         const amount = dict[sku];
-        const name = schema
-            .getName(SKU.fromString(sku), false)
-            .replace(/Non-Craftable/g, 'NC')
-            .replace(/Professional Killstreak/g, 'Pro KS')
-            .replace(/Specialized Killstreak/g, 'Spec KS')
-            .replace(/Killstreak/g, 'KS');
+        const name = replaceItemName(schema.getName(SKU.fromString(sku), false));
 
         summary.push('[' + name + '](https://www.prices.tf/items/' + sku + ')' + (amount > 1 ? ' x' + amount : ''));
     }
